refactor(dashboard): split approval review workspace
