+++ dashboard/src/review-workspace.tsx
@@ -1,79 +1,12 @@
-import { forwardRef, useCallback, useEffect, useMemo, useRef, useState, type ChangeEvent } from "react";
-import {
-  HiMiniCheckCircle,
-  HiMiniNoSymbol,
-  HiMiniShieldCheck,
-  HiMiniExclamationTriangle,
-  HiMiniInformationCircle,
-  HiMiniChevronLeft,
-  HiMiniDocumentText,
-  HiMiniArrowPath,
-  HiMiniChevronDown,
-  HiMiniChevronUp,
-  HiMiniArrowTopRightOnSquare,
-  HiMiniClock,
-  HiMiniClipboardDocumentCheck,
-  HiMiniCodeBracket,
-  HiMiniCommandLine,
-  HiMiniCog6Tooth,
-  HiMiniCube,
-  HiMiniDocumentMagnifyingGlass,
-  HiMiniDocumentPlus,
-  HiMiniGlobeAlt,
-  HiMiniKey,
-  HiMiniPencilSquare,
-  HiMiniServerStack,
-} from "react-icons/hi2";
-import {
-  ActionButton,
-  Badge,
-  EmptyState,
-  SectionLabel,
-  Tag,
-  GuardHero,
-  ProofStrip,
-} from "./approval-center-primitives";
-import {
-  buildRetryAfterApprovalCopy,
-  buildPrimaryReviewAction,
-  resolveSecondaryRiskSummary,
-  harnessDisplayName,
-  displayArtifactName,
-  formatRelativeTime,
-  buildCodexResumeUx,
-} from "./approval-center-utils";
-import {
-  SkillRiskCard,
-  SupplyChainRiskCard,
-  DecodedLayerCard,
-} from "./risk-signal-cards";
-import { DataFlowEvidenceCard } from "./data-flow-evidence-card";
-import { ScannerEvidenceSection } from "./scanner-evidence-badge";
-import {
-  advancedScopeChoicesForRequest,
-  buildDecisionPayload,
-  recommendedScopeForAction,
-  scopeChoicesForRequest,
-  standardScopeChoicesForRequest,
-  taskCapabilityExplanation,
-  type ApprovalScopeChoice,
-} from "./approval-scopes";
-import { ConsolidatedEvidenceAlert, type EvidenceItem } from "./consolidated-evidence-alert";
-import { protectionHealthFor, unavailableProtectionHealth } from "./protection-health";
-import { useRequestReadState, type RequestReadState, REQUEST_READ_STATE_LIMIT } from "./request-read-state";
-import {
-  deriveDataFlowEvidence,
-  deriveSkillRiskSignals,
-  deriveSupplyChainRiskSignals,
-  deriveEncodedLayerSignals,
-} from "./approval-center-utils";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { HiMiniChevronDown } from "react-icons/hi2";
+import { useRequestReadState } from "./request-read-state";
 import type {
   GuardApprovalGatePublicConfig,
   GuardApprovalRequest,
   GuardArtifactDiff,
   GuardCodexResumeResult,
   GuardPolicyDecision,
-  GuardProtectionState,
   GuardReceipt,
   GuardRuntimeSnapshot,
   DecisionScope,
@@ -93,12 +26,8 @@ import {
   type QueueSortDirection,
   type SemanticGroupId,
 } from "./queue-state";
-import { plainEnglishRequestTitle, whyPaused } from "./evidence/plain-english";
-import { requiresApprovalPasswordPrompt, type BulkGateCredentials } from "./approval-gate-utils";
-import { ApprovalPasswordModal } from "./approval-center-review-cards";
-import { approvalProofRequiresPassword } from "./approval-proof-inline";
+import type { BulkGateCredentials } from "./approval-gate-utils";
 import { guardAwareHref } from "./guard-api";
-import { LoggedActionPanel } from "./logged-action-panel";
 import {
   QueueBulkDrawer,
   QueueBulkGatePrompt,
@@ -106,6 +35,9 @@ import {
   QueueBulkStatusBanner,
 } from "./queue-bulk-approve-flow";
 import { useQueueBulkApprove } from "./use-queue-bulk-approve";
+import { ReviewDecisionCard } from "./review-decision-card";
+import { ReviewHeader, ReviewQueueList } from "./review-queue-list";
+import { ReviewEmptyState } from "./review-states";
 
 export type ReviewViewModel = {
   item: GuardApprovalRequest;
@@ -114,29 +46,7 @@ export type ReviewViewModel = {
   policy: GuardPolicyDecision[];
 };
 
-const PROTECTION_APPEARANCE = {
-  protected: {
-    Icon: HiMiniShieldCheck,
-    cardClass: "border-emerald-200/60 bg-emerald-50/30",
-    iconClass: "bg-brand-green/10 text-brand-green",
-  },
-  partial: {
-    Icon: HiMiniInformationCircle,
-    cardClass: "border-brand-blue/20 bg-brand-blue/[0.04]",
-    iconClass: "bg-brand-blue/10 text-brand-blue",
-  },
-  degraded: {
-    Icon: HiMiniExclamationTriangle,
-    cardClass: "border-brand-attention/20 bg-brand-attention/[0.04]",
-    iconClass: "bg-brand-attention/10 text-brand-attention",
-  },
-} satisfies Record<GuardProtectionState, {
-  Icon: typeof HiMiniShieldCheck;
-  cardClass: string;
-  iconClass: string;
-}>;
-
-type ReviewWorkspaceProps = {
+export type ReviewWorkspaceProps = {
   requests: GuardApprovalRequest[];
   activeRequestId: string | null;
   detail: ReviewViewModel | null;
@@ -163,8 +73,6 @@ type ReviewWorkspaceProps = {
 };
 
 const QUEUE_PAGE_SIZE = 10;
-const commonScopeValues = new Set<DecisionScope>(["artifact"]);
-
 export function ReviewWorkspace(props: ReviewWorkspaceProps) {
   const { requests, activeRequestId, detail } = props;
   const readState = useRequestReadState();
@@ -423,1468 +331,4 @@ export function ReviewWorkspace(props: ReviewWorkspaceProps) {
       </div>
     </div>
   );
-}
-
-function ReviewHeader({
-  count,
-  filteredCount,
-  progress,
-  activeHarness,
-}: {
-  count: number;
-  filteredCount: number;
-  progress: string;
-  activeHarness: string;
-}) {
-  const isFiltered = filteredCount !== count;
-  return (
-    <div className="flex flex-col gap-3 lg:flex-row lg:items-center lg:justify-between">
-      <div>
-        <h1 className="text-xl font-semibold tracking-[-0.02em] text-brand-dark sm:text-2xl">Review</h1>
-        <p className="mt-1 max-w-xl text-sm leading-relaxed text-muted-foreground">
-          Guard paused these actions before they ran. Review each one and decide what should happen.
-        </p>
-      </div>
-      <p className="text-sm text-muted-foreground">
-        {progress}{isFiltered ? ` · ${filteredCount} of ${count} shown` : ""} · from {harnessDisplayName(activeHarness)}
-      </p>
-    </div>
-  );
-}
-
-function resolveSemanticGroup(categoryId: QueueCategoryId): SemanticGroupId {
-  for (const group of REVIEW_SEMANTIC_GROUPS) {
-    if (group.matches.includes(categoryId)) return group.id;
-  }
-  return "other";
-}
-
-const ReviewQueueList = forwardRef<HTMLDivElement, {
-  requests: GuardApprovalRequest[];
-  allFilteredRequests: GuardApprovalRequest[];
-  totalCount: number;
-  filteredCount: number;
-  activeRequestId: string | null;
-  readState: RequestReadState;
-  categoryOptions: QueueCategory[];
-  categoryFilter: QueueCategoryId | "all";
-  searchTerm: string;
-  sortDirection: QueueSortDirection;
-  semanticFilter: SemanticGroupId;
-  dateFrom: string;
-  dateTo: string;
-  page: number;
-  totalPages: number;
-  onCategoryFilterChange: (category: QueueCategoryId | "all") => void;
-  onSearchTermChange: (term: string) => void;
-  onSortDirectionChange: (direction: QueueSortDirection) => void;
-  onSemanticFilterChange: (group: SemanticGroupId) => void;
-  onDateFromChange: (date: string) => void;
-  onDateToChange: (date: string) => void;
-  onPageChange: (page: number) => void;
-  onOpenRequest: (requestId: string) => void;
-  selectionMode?: boolean;
-  isBulkSelectable?: (item: GuardApprovalRequest) => boolean;
-  isBulkSelected?: (item: GuardApprovalRequest) => boolean;
-  onBulkToggleSelect?: (item: GuardApprovalRequest) => void;
-  onBulkSelectAll?: () => void;
-  onBulkClearAll?: () => void;
-}>(({
-  requests,
-  allFilteredRequests,
-  totalCount,
-  filteredCount,
-  activeRequestId,
-  readState,
-  categoryOptions,
-  categoryFilter,
-  searchTerm,
-  sortDirection,
-  semanticFilter,
-  dateFrom,
-  dateTo,
-  page,
-  totalPages,
-  onCategoryFilterChange,
-  onSearchTermChange,
-  onSortDirectionChange,
-  onSemanticFilterChange,
-  onDateFromChange,
-  onDateToChange,
-  onPageChange,
-  onOpenRequest,
-  selectionMode = false,
-  isBulkSelectable,
-  isBulkSelected,
-  onBulkToggleSelect,
-  onBulkSelectAll,
-  onBulkClearAll,
-}, ref) => {
-  const [showFilters, setShowFilters] = useState(false);
-
-  const handleSearchChange = useCallback((event: ChangeEvent<HTMLInputElement>) => {
-    onSearchTermChange(event.target.value);
-  }, [onSearchTermChange]);
-  const handleCategoryChange = useCallback((event: ChangeEvent<HTMLSelectElement>) => {
-    onCategoryFilterChange(event.target.value as QueueCategoryId | "all");
-  }, [onCategoryFilterChange]);
-  const handleSortChange = useCallback((event: ChangeEvent<HTMLSelectElement>) => {
-    onSortDirectionChange(event.target.value as QueueSortDirection);
-  }, [onSortDirectionChange]);
-  const handleDateFromChange = useCallback((event: ChangeEvent<HTMLInputElement>) => {
-    onDateFromChange(event.target.value);
-  }, [onDateFromChange]);
-  const handleDateToChange = useCallback((event: ChangeEvent<HTMLInputElement>) => {
-    onDateToChange(event.target.value);
-  }, [onDateToChange]);
-  const handleToggleFilters = useCallback(() => {
-    setShowFilters((visible) => !visible);
-  }, []);
-  const handleClearFilters = useCallback(() => {
-    onSearchTermChange("");
-    onCategoryFilterChange("all");
-    onSortDirectionChange("newest");
-    onSemanticFilterChange("all");
-    onDateFromChange("");
-    onDateToChange("");
-  }, [
-    onCategoryFilterChange,
-    onDateFromChange,
-    onDateToChange,
-    onSearchTermChange,
-    onSemanticFilterChange,
-    onSortDirectionChange,
-  ]);
-  const handlePreviousPage = useCallback(() => {
-    onPageChange(Math.max(1, page - 1));
-  }, [page, onPageChange]);
-  const handleNextPage = useCallback(() => {
-    onPageChange(Math.min(totalPages, page + 1));
-  }, [page, totalPages, onPageChange]);
-
-  const activeSemanticGroup = semanticFilter;
-
-  const visibleGroups = useMemo(() => {
-    const available = new Set<SemanticGroupId>();
-    for (const item of allFilteredRequests) {
-      available.add(resolveSemanticGroup(resolveQueueCategory(item).id));
-    }
-    return REVIEW_SEMANTIC_GROUPS.filter((g) => g.id === "all" || available.has(g.id));
-  }, [allFilteredRequests]);
-
-  const isFiltered =
-    searchTerm ||
-    semanticFilter !== "all" ||
-    categoryFilter !== "all" ||
-    sortDirection !== "newest" ||
-    dateFrom ||
-    dateTo;
-  const showPagination = filteredCount > QUEUE_PAGE_SIZE;
-
-  // Page-level select-all state for the compact list header. Reflects only the
-  // currently visible rows so the checkbox behaves like a standard list header
-  // (indeterminate when some-but-not-all eligible rows on the page are selected).
-  const pageSelectableItems = useMemo(
-    () => requests.filter((item) => isBulkSelectable?.(item) === true),
-    [requests, isBulkSelectable],
-  );
-  const pageSelectedCount = useMemo(
-    () => pageSelectableItems.filter((item) => isBulkSelected?.(item) === true).length,
-    [pageSelectableItems, isBulkSelected],
-  );
-  const pageAllSelected =
-    pageSelectableItems.length > 0 && pageSelectedCount === pageSelectableItems.length;
-  const pageSomeSelected = pageSelectedCount > 0 && !pageAllSelected;
-
-  const handlePageSelectAll = useCallback(() => {
-    if (pageAllSelected) {
-      onBulkClearAll?.();
-    } else {
-      onBulkSelectAll?.();
-    }
-  }, [pageAllSelected, onBulkSelectAll, onBulkClearAll]);
-
-  // Stable ref callback so React does not detach/reattach it every render.
-  const setIndeterminate = useCallback(
-    (el: HTMLInputElement | null) => {
-      if (el) el.indeterminate = pageSomeSelected;
-    },
-    [pageSomeSelected],
-  );
-
-  return (
-    <aside className="space-y-3" ref={ref}>
-      <div className="flex items-center justify-between gap-3">
-        <SectionLabel>Queue</SectionLabel>
-        <div className="flex items-center gap-3">
-          <button
-            type="button"
-            disabled={
-              allFilteredRequests.length > 0 &&
-              allFilteredRequests.length +
-                readState.readCount -
-                allFilteredRequests.filter((item) => readState.isRead(item.request_id)).length >
-                REQUEST_READ_STATE_LIMIT
-            }
-            title={
-              allFilteredRequests.length +
-                readState.readCount -
-                allFilteredRequests.filter((item) => readState.isRead(item.request_id)).length >
-                REQUEST_READ_STATE_LIMIT
-                ? `Cannot mark all read: doing so would exceed the read-state storage cap (${REQUEST_READ_STATE_LIMIT.toLocaleString()}). Reduce filters to shrink the visible queue, or mark requests read one by one.`
-                : `Marks every visible filtered request as read (remembering up to ${REQUEST_READ_STATE_LIMIT.toLocaleString()}).`
-            }
-            onClick={() => readState.markAllRead(allFilteredRequests.map((item) => item.request_id))}
-            className={`text-xs font-medium transition-colors ${
-              allFilteredRequests.length +
-                readState.readCount -
-                allFilteredRequests.filter((item) => readState.isRead(item.request_id)).length >
-                REQUEST_READ_STATE_LIMIT
-                ? "text-slate-400 cursor-not-allowed"
-                : "text-brand-blue hover:text-brand-dark"
-            }`}
-          >
-            Mark all read
-          </button>
-          <span className="font-mono text-[11px] font-semibold text-muted-foreground">
-            {filteredCount}/{totalCount}
-          </span>
-        </div>
-      </div>
-      <div className="space-y-2 rounded-xl border border-slate-100 bg-white p-3">
-        <label className="block">
-          <span className="sr-only">Search review queue</span>
-          <input
-            id="guard-review-queue-search"
-            name="guard-review-queue-search"
-            type="search"
-            value={searchTerm}
-            onChange={handleSearchChange}
-            placeholder="Search queue..."
-            className="min-h-10 w-full rounded-lg border border-slate-200 bg-white px-3 text-sm text-brand-dark placeholder:text-slate-400 focus:border-brand-blue focus:outline-none focus:ring-2 focus:ring-brand-blue/20"
-          />
-        </label>
-        <button
-          type="button"
-          onClick={handleToggleFilters}
-          className="flex items-center gap-1 text-xs font-medium text-brand-blue hover:text-brand-dark transition-colors"
-        >
-          {showFilters ? "Hide filters" : "Show filters"}
-          {isFiltered && !showFilters && <span className="ml-1 h-1.5 w-1.5 rounded-full bg-brand-attention" />}
-        </button>
-        {showFilters && (
-          <div className="space-y-2">
-            <div className="flex flex-wrap gap-1">
-              {visibleGroups.map((group) => (
-                <SemanticFilterButton
-                  key={group.id}
-                  group={group}
-                  selected={activeSemanticGroup === group.id}
-                  onSelect={onSemanticFilterChange}
-                />
-              ))}
-            </div>
-            <label className="block">
-              <span className="mb-1 block text-[11px] font-semibold uppercase tracking-[0.16em] text-muted-foreground">
-                Sort by date
-              </span>
-              <select
-                value={sortDirection}
-                onChange={handleSortChange}
-                aria-label="Sort review queue"
-                className="min-h-10 w-full rounded-lg border border-slate-200 bg-white px-3 text-sm font-medium text-brand-dark focus:border-brand-blue focus:outline-none focus:ring-2 focus:ring-brand-blue/20"
-              >
-                <option value="newest">Newest first</option>
-                <option value="oldest">Oldest first</option>
-                <option value="highest_risk">Highest risk first</option>
-                <option value="category">Category</option>
-              </select>
-            </label>
-            <div className="grid gap-2 sm:grid-cols-2">
-              <label className="block">
-                <span className="mb-1 block text-[11px] font-semibold uppercase tracking-[0.16em] text-muted-foreground">
-                  From date
-                </span>
-                <input
-                  type="date"
-                  value={dateFrom}
-                  onChange={handleDateFromChange}
-                  aria-label="Filter requests from date"
-                  className="min-h-10 w-full rounded-lg border border-slate-200 bg-white px-3 text-sm font-medium text-brand-dark focus:border-brand-blue focus:outline-none focus:ring-2 focus:ring-brand-blue/20"
-                />
-              </label>
-              <label className="block">
-                <span className="mb-1 block text-[11px] font-semibold uppercase tracking-[0.16em] text-muted-foreground">
-                  To date
-                </span>
-                <input
-                  type="date"
-                  value={dateTo}
-                  onChange={handleDateToChange}
-                  aria-label="Filter requests to date"
-                  className="min-h-10 w-full rounded-lg border border-slate-200 bg-white px-3 text-sm font-medium text-brand-dark focus:border-brand-blue focus:outline-none focus:ring-2 focus:ring-brand-blue/20"
-                />
-              </label>
-            </div>
-            {isFiltered && (
-              <button
-                type="button"
-                onClick={handleClearFilters}
-                className="text-xs font-medium text-brand-blue hover:text-brand-dark transition-colors"
-              >
-                Clear all filters
-              </button>
-            )}
-          </div>
-        )}
-      </div>
-      <div
-        role="listbox"
-        aria-label="Review queue"
-        className="space-y-2 rounded-lg border border-slate-100 bg-white p-1.5"
-      >
-        {selectionMode && pageSelectableItems.length > 0 && (
-          <div className="flex items-center gap-2 border-b border-slate-100 px-2 pb-1.5 pt-1">
-            <label className="flex shrink-0 cursor-pointer items-center gap-2">
-              <input
-                type="checkbox"
-                checked={pageAllSelected}
-                ref={setIndeterminate}
-                onChange={handlePageSelectAll}
-                aria-label={
-                  pageAllSelected
-                    ? "Clear selection of eligible reads on this page"
-                    : "Select all eligible reads on this page"
-                }
-                className="h-4 w-4 rounded border-slate-300 text-brand-blue focus:ring-brand-blue/30"
-              />
-              <span className="text-[11px] font-medium text-muted-foreground">
-                {pageSelectedCount > 0
-                  ? `${pageSelectedCount} of ${pageSelectableItems.length} selected`
-                  : `Select all eligible (${pageSelectableItems.length})`}
-              </span>
-            </label>
-          </div>
-        )}
-        {requests.length > 0 ? (
-          requests.map((item, index) => (
-            <QueueItemRow
-              key={item.request_id}
-              item={item}
-              active={item.request_id === activeRequestId}
-              readState={readState}
-              index={index}
-              onOpenRequest={onOpenRequest}
-              selectionMode={selectionMode}
-              selectable={isBulkSelectable?.(item) === true}
-              selected={isBulkSelected?.(item) === true}
-              onToggleSelect={onBulkToggleSelect}
-            />
-          ))
-        ) : (
-          <div className="px-3 py-5">
-            <EmptyState title="No matching actions" body="Try a different search or filter." tone="teach" />
-          </div>
-        )}
-      </div>
-      {showPagination && (
-        <div className="flex items-center justify-between gap-2 text-xs text-muted-foreground">
-          <span>
-            {((page - 1) * QUEUE_PAGE_SIZE) + 1}-{Math.min(filteredCount, page * QUEUE_PAGE_SIZE)} of {filteredCount}
-          </span>
-          <div className="flex items-center gap-2">
-            <button
-              type="button"
-              onClick={handlePreviousPage}
-              disabled={page <= 1}
-              className="min-h-9 rounded-lg border border-slate-200 bg-white px-3 font-semibold text-brand-dark transition-colors duration-150 hover:border-brand-blue/30 disabled:pointer-events-none disabled:opacity-40"
-            >
-              Previous
-            </button>
-            <span className="font-mono text-[11px] text-slate-400">
-              {page}/{totalPages}
-            </span>
-            <button
-              type="button"
-              onClick={handleNextPage}
-              disabled={page >= totalPages}
-              className="min-h-9 rounded-lg border border-slate-200 bg-white px-3 font-semibold text-brand-dark transition-colors duration-150 hover:border-brand-blue/30 disabled:pointer-events-none disabled:opacity-40"
-            >
-              Next
-            </button>
-          </div>
-        </div>
-      )}
-    </aside>
-  );
-});
-ReviewQueueList.displayName = "ReviewQueueList";
-
-function SemanticFilterButton(props: {
-  group: (typeof REVIEW_SEMANTIC_GROUPS)[number];
-  selected: boolean;
-  onSelect: (group: SemanticGroupId) => void;
-}) {
-  const { group, selected, onSelect } = props;
-  const handleSelect = useCallback(() => {
-    onSelect(group.id);
-  }, [group.id, onSelect]);
-
-  return (
-    <button
-      type="button"
-      onClick={handleSelect}
-      className={`rounded-full px-2.5 py-1 text-[11px] font-medium transition-all ${
-        selected
-          ? "bg-brand-blue text-white"
-          : "border border-slate-200 bg-white text-brand-dark hover:bg-slate-50"
-      }`}
-    >
-      {group.label}
-    </button>
-  );
-}
-
-type RiskLevel = "high" | "medium" | "low";
-
-function riskLevelFromScore(score: number): RiskLevel {
-  if (score <= 2) return "high";
-  if (score <= 4) return "medium";
-  return "low";
-}
-function QueueItemRow({ item, active, readState, index, onOpenRequest, selectionMode = false, selectable = false, selected = false, onToggleSelect }: {
-  item: GuardApprovalRequest;
-  active: boolean;
-  readState: RequestReadState;
-  index: number;
-  onOpenRequest: (requestId: string) => void;
-  selectionMode?: boolean;
-  selectable?: boolean;
-  selected?: boolean;
-  onToggleSelect?: (item: GuardApprovalRequest) => void;
-}) {
-  const risk = riskScore(item);
-  const riskLevel = riskLevelFromScore(risk);
-  const category = resolveQueueCategory(item);
-  const CategoryIcon = iconForQueueCategory(category.id);
-  const preview = queueItemPreview(item);
-  const isRead = readState.isRead(item.request_id);
-  // Checkboxes render whenever bulk selection is active so the affordance is
-  // always discoverable. Non-eligible rows show a disabled checkbox with a
-  // tooltip instead of silently hiding the control.
-  const showCheckbox = selectionMode;
-  const canSelect = selectionMode && selectable;
-
-  const handleClick = useCallback(() => {
-    onOpenRequest(item.request_id);
-  }, [item.request_id, onOpenRequest]);
-
-  const handleCheckboxChange = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => {
-      event.stopPropagation();
-      if (!canSelect) return;
-      onToggleSelect?.(item);
-    },
-    [item, onToggleSelect, canSelect],
-  );
-
-  const handleKeyDown = useCallback(
-    (event: React.KeyboardEvent<HTMLLabelElement>) => {
-      if (!canSelect) return;
-      // Space is handled natively by the checkbox itself; only remap Enter
-      // here so the wrapping label is also keyboard-activatable.
-      if (event.key === "Enter") {
-        event.preventDefault();
-        event.stopPropagation();
-        onToggleSelect?.(item);
-      }
-    },
-    [item, onToggleSelect, canSelect],
-  );
-
-  const checkboxLabel = canSelect
-    ? `Select ${preview} for bulk approval`
-    : `Not eligible for bulk approval: ${category.shortLabel.toLowerCase()}`;
-
-  const rowClassName = (() => {
-    if (selected) return "border border-brand-blue/60 bg-brand-blue/[0.08] ring-1 ring-brand-blue/20";
-    if (active) return "border border-brand-blue bg-brand-blue/[0.06]";
-    if (isRead) return "border border-transparent bg-white hover:bg-slate-50";
-    return "border border-transparent bg-slate-50 hover:bg-slate-100";
-  })();
-
-  return (
-    <div
-      role="none"
-      className={`group w-full rounded-lg py-2.5 px-2 transition-all ${rowClassName}`}
-    >
-      <div className="flex items-center justify-between gap-2">
-        {showCheckbox ? (
-          <label
-            className={`flex shrink-0 items-center ${canSelect ? "cursor-pointer" : "cursor-not-allowed"}`}
-            title={checkboxLabel}
-            onClick={(event) => event.stopPropagation()}
-            onKeyDown={handleKeyDown}
-          >
-            <input
-              type="checkbox"
-              checked={selected}
-              disabled={!canSelect}
-              onChange={handleCheckboxChange}
-              aria-label={checkboxLabel}
-              className="h-4 w-4 rounded border-slate-300 text-brand-blue focus:ring-brand-blue/30 disabled:opacity-40"
-            />
-          </label>
-        ) : null}
-        <button
-          type="button"
-          onClick={handleClick}
-          role="option"
-          aria-selected={active}
-          aria-posinset={index + 1}
-          aria-setsize={undefined}
-          tabIndex={active ? 0 : -1}
-          className="flex min-w-0 flex-1 items-center gap-2 text-left"
-        >
-          <div className="min-w-0 flex-1">
-            <p className={`truncate text-sm ${isRead ? "font-medium text-slate-500" : "font-bold text-brand-dark"}`}>
-              {!isRead && <span className="sr-only">Unread request:</span>}
-              {preview}
-            </p>
-            <p className="truncate text-[11px] text-muted-foreground">
-              {harnessDisplayName(item.harness)} · {category.shortLabel} · {formatQueueRequestDate(item)}
-            </p>
-          </div>
-          <span
-            role="img"
-            aria-label={`Risk: ${riskLevel}`}
-            className="group/icon relative flex h-2 w-2 shrink-0 items-center justify-center"
-          >
-            <span
-              className={`h-2 w-2 rounded-full ${
-                riskLevel === "high" ? "bg-red-400" : riskLevel === "medium" ? "bg-amber-400" : "bg-emerald-400"
-              }`}
-            />
-            <span className="pointer-events-none absolute right-0 top-full z-50 mt-1.5 whitespace-nowrap rounded-md bg-brand-blue px-2 py-1 text-[10px] font-medium text-white opacity-0 shadow-lg transition-opacity duration-150 group-hover/icon:opacity-100">
-              {`Risk: ${riskLevel}`}
-            </span>
-          </span>
-          <span
-            role="img"
-            aria-label={category.label}
-            className={`group/icon relative inline-flex h-7 w-7 shrink-0 items-center justify-center rounded-lg ${
-              active ? "bg-brand-blue/10 text-brand-blue" : "bg-slate-50 text-slate-500"
-            }`}
-          >
-            <CategoryIcon className="h-4 w-4" aria-hidden="true" />
-            <span className="pointer-events-none absolute right-0 top-full z-50 mt-1.5 whitespace-nowrap rounded-md bg-brand-blue px-2 py-1 text-[10px] font-medium text-white opacity-0 shadow-lg transition-opacity duration-150 group-hover/icon:opacity-100">
-              {category.label}
-            </span>
-          </span>
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function iconForQueueCategory(categoryId: QueueCategoryId) {
-  switch (categoryId) {
-    case "credential_output":
-      return HiMiniKey;
-    case "secret_file_read":
-      return HiMiniDocumentMagnifyingGlass;
-    case "file_read":
-      return HiMiniDocumentMagnifyingGlass;
-    case "secret_exfiltration":
-      return HiMiniArrowTopRightOnSquare;
-    case "system_prompt_access":
-      return HiMiniInformationCircle;
-    case "prompt_injection":
-      return HiMiniExclamationTriangle;
-    case "guard_bypass":
-      return HiMiniNoSymbol;
-    case "generated_inventory_edit":
-      return HiMiniClipboardDocumentCheck;
-    case "docs_edit":
-      return HiMiniDocumentText;
-    case "source_edit":
-      return HiMiniPencilSquare;
-    case "config_change":
-      return HiMiniCog6Tooth;
-    case "file_upload":
-      return HiMiniArrowTopRightOnSquare;
-    case "file_delete_cleanup":
-      return HiMiniNoSymbol;
-    case "git_operation":
-      return HiMiniCodeBracket;
-    case "process_control":
-      return HiMiniArrowPath;
-    case "container_or_deploy":
-      return HiMiniServerStack;
-    case "persistence_change":
-      return HiMiniClock;
-    case "package_install":
-      return HiMiniCube;
-    case "package_script":
-      return HiMiniCommandLine;
-    case "destructive_shell":
-      return HiMiniNoSymbol;
-    case "encoded_shell":
-      return HiMiniCodeBracket;
-    case "network":
-      return HiMiniGlobeAlt;
-    case "mcp_tool":
-      return HiMiniServerStack;
-    case "browser_action":
-      return HiMiniArrowTopRightOnSquare;
-    case "harness_start":
-      return HiMiniShieldCheck;
-    case "shell_command":
-      return HiMiniCommandLine;
-    case "other":
-      return HiMiniDocumentPlus;
-  }
-}
-
-function queueItemPreview(item: GuardApprovalRequest): string {
-  const envelope = item.action_envelope_json;
-  return (
-    envelope?.command ??
-    item.raw_command_text ??
-    envelope?.mcp_tool ??
-    (envelope?.prompt_text ?? envelope?.prompt_excerpt) ??
-    envelope?.package_name ??
-    displayArtifactName(item)
-  );
-}
-
-function ReviewDecisionCard(props: {
-  detail: ReviewViewModel | null;
-  onResolve: ReviewWorkspaceProps["onResolve"];
-  onGoHome: () => void;
-  approvalGate: GuardApprovalGatePublicConfig | null;
-}) {
-  const detail = props.detail;
-  const item = detail?.item ?? null;
-  const [allowScope, setAllowScope] = useState<DecisionScope>("artifact");
-  const [blockScope, setBlockScope] = useState<DecisionScope>("artifact");
-  const [submitting, setSubmitting] = useState<"allow" | "block" | null>(null);
-  const [resolved, setResolved] = useState<"allow" | "block" | null>(null);
-  const [showConsequences, setShowConsequences] = useState(false);
-  const [showEvidence, setShowEvidence] = useState(false);
-  const [lastAction, setLastAction] = useState<"allow" | "block" | null>(null);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
-  const [approvalPassword, setApprovalPassword] = useState("");
-  const [approvalTotpCode, setApprovalTotpCode] = useState("");
-  const [useCooldown, setUseCooldown] = useState(false);
-  const [pendingAction, setPendingAction] = useState<"allow" | "block" | null>(null);
-  const [pendingContractKey, setPendingContractKey] = useState<string | null>(null);
-  const timerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const allowButtonRef = useRef<HTMLButtonElement>(null);
-  const availableScopeChoices = useMemo(
-    () => (item ? standardScopeChoicesForRequest(item, "allow") : []),
-    [item]
-  );
-  const commonScopeOptions = useMemo(
-    () => availableScopeChoices.filter((choice) => commonScopeValues.has(choice.value)),
-    [availableScopeChoices]
-  );
-  const broaderScopeOptions = useMemo(
-    () => availableScopeChoices.filter((choice) => !commonScopeValues.has(choice.value)),
-    [availableScopeChoices]
-  );
-  const advancedScopeOptions = useMemo(
-    () => (item ? advancedScopeChoicesForRequest(item, "allow") : []),
-    [item]
-  );
-  const blockScopeOptions = useMemo(
-    () => (item ? scopeChoicesForRequest(item, "block") : []),
-    [item],
-  );
-  const taskCapabilityCopy = item ? taskCapabilityExplanation(item) : null;
-  const hasAllowScope = availableScopeChoices.length + advancedScopeOptions.length > 0;
-  const decisionContractKey = item
-    ? `${item.request_id}:${item.scope_contract_version ?? "legacy"}:${item.scope_contract_digest ?? "legacy"}`
-    : null;
-
-  useEffect(() => {
-    if (item) {
-      setAllowScope(recommendedScopeForAction(item, "allow") ?? "artifact");
-      setBlockScope(recommendedScopeForAction(item, "block") ?? "artifact");
-      setResolved(null);
-      setSubmitting(null);
-      setLastAction(null);
-      setErrorMessage(null);
-      setApprovalPassword("");
-      setApprovalTotpCode("");
-      setUseCooldown(false);
-      setPendingAction(null);
-      setPendingContractKey(null);
-    }
-  }, [item?.request_id, item?.scope_contract_version, item?.scope_contract_digest]);
-
-  useEffect(() => {
-    return () => {
-      if (timerRef.current) clearTimeout(timerRef.current);
-    };
-  }, []);
-
-  const handleResolve = useCallback(
-    async (action: "allow" | "block") => {
-      if (!item) return;
-      setSubmitting(action);
-      setErrorMessage(null);
-      try {
-        const requestedScope = action === "allow" ? allowScope : blockScope;
-        const gate = props.approvalGate;
-        const needsPassword = approvalProofRequiresPassword(gate);
-        const includeGateFields =
-          gate?.enabled === true &&
-          gate?.configured === true &&
-          requiresApprovalPasswordPrompt(gate.cooldown_active, gate.strict_all_decisions, requestedScope);
-        await props.onResolve({
-          ...buildDecisionPayload({
-            item,
-            action,
-            scope: requestedScope,
-            reason: action === "allow" ? "approved in review" : "blocked in review",
-          }),
-          ...(includeGateFields && needsPassword ? { approval_password: approvalPassword } : {}),
-          ...(includeGateFields && !needsPassword ? { approval_totp_code: approvalTotpCode } : {}),
-          ...(includeGateFields ? { approval_gate_use_cooldown: useCooldown } : {}),
-        });
-        setResolved(action);
-        setApprovalPassword("");
-        setApprovalTotpCode("");
-        setUseCooldown(false);
-        setPendingAction(null);
-        setPendingContractKey(null);
-        timerRef.current = setTimeout(() => setResolved(null), 2000);
-      } catch (err) {
-        setErrorMessage(err instanceof Error ? err.message : "Something went wrong. Try again.");
-      } finally {
-        setSubmitting(null);
-      }
-    },
-    [
-      item,
-      allowScope,
-      blockScope,
-      props.onResolve,
-      props.approvalGate,
-      approvalPassword,
-      approvalTotpCode,
-      useCooldown,
-    ]
-  );
-
-  const handleRequestResolve = useCallback(
-    (action: "allow" | "block") => {
-      if (action === "allow" && !hasAllowScope) {
-        setErrorMessage("This action has no eligible approval scope.");
-        return;
-      }
-      if (action === "block" && blockScopeOptions.length === 0) {
-        setErrorMessage("This action has no eligible block scope.");
-        return;
-      }
-      setLastAction(action);
-      const requestedScope = action === "allow" ? allowScope : blockScope;
-      const gate = props.approvalGate;
-      const gateRequiresPassword =
-        gate?.enabled === true &&
-        gate?.configured === true &&
-        requiresApprovalPasswordPrompt(gate.cooldown_active, gate.strict_all_decisions, requestedScope);
-      if (gateRequiresPassword) {
-        setPendingAction(action);
-        setPendingContractKey(decisionContractKey);
-        setErrorMessage(null);
-        return;
-      }
-      void handleResolve(action);
-    },
-    [
-      allowScope,
-      blockScope,
-      blockScopeOptions.length,
-      decisionContractKey,
-      handleResolve,
-      hasAllowScope,
-      props.approvalGate,
-    ]
-  );
-
-  const handleAllow = useCallback(() => {
-    handleRequestResolve("allow");
-  }, [handleRequestResolve]);
-  const handleBlock = useCallback(() => {
-    handleRequestResolve("block");
-  }, [handleRequestResolve]);
-
-  useEffect(() => {
-    function handleKeyDown(event: KeyboardEvent) {
-      if (submitting !== null || pendingAction !== null) return;
-      const target = event.target as HTMLElement;
-      if (target.tagName === "INPUT" || target.tagName === "TEXTAREA" || target.isContentEditable) return;
-
-      if (event.key === "a" || event.key === "A") {
-        event.preventDefault();
-        handleRequestResolve("allow");
-      }
-      if (event.key === "b" || event.key === "B") {
-        event.preventDefault();
-        handleRequestResolve("block");
-      }
-      const scopeIndex = parseInt(event.key, 10);
-      if (scopeIndex >= 1 && scopeIndex <= availableScopeChoices.length) {
-        event.preventDefault();
-        setAllowScope(availableScopeChoices[scopeIndex - 1].value);
-      }
-    }
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [availableScopeChoices, handleRequestResolve, pendingAction, submitting]);
-
-  const handleModalSubmit = useCallback(() => {
-    if (pendingAction === null) {
-      return;
-    }
-    if (pendingContractKey !== decisionContractKey) {
-      setPendingAction(null);
-      setPendingContractKey(null);
-      setErrorMessage("This request changed while you were reviewing it. Review the current scopes and try again.");
-      return;
-    }
-    void handleResolve(pendingAction);
-  }, [decisionContractKey, handleResolve, pendingAction, pendingContractKey]);
-
-  const handleModalCancel = useCallback(() => {
-    setPendingAction(null);
-    setPendingContractKey(null);
-    setApprovalPassword("");
-    setApprovalTotpCode("");
-    setUseCooldown(false);
-  }, []);
-
-  const handleToggleConsequences = useCallback(() => {
-    setShowConsequences((visible) => !visible);
-  }, []);
-  const handleToggleEvidence = useCallback(() => {
-    setShowEvidence((visible) => !visible);
-  }, []);
-  const handleRetryLastAction = useCallback(() => {
-    setErrorMessage(null);
-    if (lastAction !== null) {
-      handleRequestResolve(lastAction);
-    }
-  }, [handleRequestResolve, lastAction]);
-
-  const handleApprovalPasswordChange = useCallback((event: ChangeEvent<HTMLInputElement>) => {
-    setApprovalPassword(event.target.value);
-  }, []);
-  const handleApprovalTotpCodeChange = useCallback((event: ChangeEvent<HTMLInputElement>) => {
-    setApprovalTotpCode(event.target.value);
-  }, []);
-
-  const handleUseCooldownChange = useCallback((event: ChangeEvent<HTMLInputElement>) => {
-    setUseCooldown(event.target.checked);
-  }, []);
-
-  if (!detail || !item) {
-    return (
-      <EmptyState
-        title="Select an action"
-        body="Choose a paused action from the queue to review and decide."
-        tone="teach"
-      />
-    );
-  }
-
-  const plainTitle = plainEnglishRequestTitle(item);
-  const harnessName = harnessDisplayName(item.harness);
-  const whatWouldHappen = buildWhatWouldHappen(item);
-  const secondaryRiskSummary = resolveSecondaryRiskSummary(item);
-  const pauseReason = whyPaused(item);
-
-  const topAlertItems: EvidenceItem[] = [];
-  if (secondaryRiskSummary) {
-    topAlertItems.push({
-      id: "secondary-risk",
-      title: "Additional risk",
-      tone: "amber",
-      icon: HiMiniExclamationTriangle,
-      content: <p className="text-sm text-brand-dark">{secondaryRiskSummary}</p>,
-    });
-  }
-  if (pauseReason) {
-    topAlertItems.push({
-      id: "why-paused",
-      title: "Why paused",
-      tone: "blue",
-      icon: HiMiniInformationCircle,
-      content: <p className="text-sm text-brand-dark">{pauseReason}</p>,
-    });
-  }
-
-  const evidenceItems: EvidenceItem[] = [];
-  const allSignals = item.decision_v2_json?.signals ?? [];
-  if (allSignals.some((s) => s.category === "skill" || s.category === "mcp")) {
-    evidenceItems.push({
-      id: "scanner",
-      title: "Scanner evidence",
-      tone: "blue",
-      content: <ScannerEvidenceSection signals={allSignals} />,
-    });
-  }
-  if (item.why_now) {
-    evidenceItems.push({
-      id: "why-now",
-      title: "Why now",
-      tone: "purple",
-      content: <p className="text-sm text-brand-dark">{item.why_now}</p>,
-    });
-  }
-  if (deriveDataFlowEvidence(item) !== null) {
-    evidenceItems.push({
-      id: "data-flow",
-      title: "Data flow detected",
-      tone: "blue",
-      content: <DataFlowEvidenceCard item={item} />,
-    });
-  }
-  if (deriveSkillRiskSignals(item).length > 0) {
-    evidenceItems.push({
-      id: "skill-risk",
-      title: "Skill risk",
-      tone: "blue",
-      content: <SkillRiskCard item={item} />,
-    });
-  }
-  const isSupplyChainArtifact =
-    item.artifact_type === "supply_chain" ||
-    item.artifact_type === "package_request" ||
-    (typeof item.artifact_type === "string" && item.artifact_type.endsWith("_package"));
-  if (deriveSupplyChainRiskSignals(item).length > 0 || isSupplyChainArtifact) {
-    evidenceItems.push({
-      id: "supply-chain",
-      title: "Supply chain risk",
-      tone: "amber",
-      content: <SupplyChainRiskCard item={item} />,
-    });
-  }
-  if (deriveEncodedLayerSignals(item).length > 0) {
-    evidenceItems.push({
-      id: "decoded-layer",
-      title: "Decoded layer",
-      tone: "slate",
-      content: <DecodedLayerCard item={item} />,
-    });
-  }
-
-  return (
-    <div className="space-y-5">
-      {resolved && (
-        <div
-          className={`guard-fade-in flex items-center gap-3 rounded-xl border px-4 py-3 transition-all ${
-            resolved === "allow"
-              ? "border-brand-green/25 bg-brand-green-bg/30"
-              : "border-brand-attention/25 bg-brand-attention/[0.04]"
-          }`}
-          role="status"
-          aria-live="polite"
-        >
-          <HiMiniCheckCircle
-            className={`h-5 w-5 shrink-0 ${resolved === "allow" ? "text-brand-green" : "text-brand-attention"}`}
-            aria-hidden="true"
-          />
-          <p className={`text-sm font-medium ${resolved === "allow" ? "text-brand-green-text" : "text-brand-attention"}`}>
-            {item ? buildRetryAfterApprovalCopy(item, resolved) : (resolved === "allow" ? "Approved: action can proceed" : "Blocked: action stopped")}
-          </p>
-        </div>
-      )}
-
-      <div className="rounded-xl border border-slate-100 p-4 sm:p-5">
-        <div className="flex items-start justify-between gap-3">
-          <div className="min-w-0 flex-1">
-            <SectionLabel>Paused action</SectionLabel>
-            <h2 className="mt-2 text-lg font-semibold text-brand-dark">{plainTitle}</h2>
-            <p className="mt-1 text-sm text-muted-foreground">
-              From {harnessName}
-            </p>
-          </div>
-          <Badge tone={item.policy_action === "block" ? "attention" : "info"}>
-            {item.policy_action === "block" ? "Blocked" : "Needs review"}
-          </Badge>
-        </div>
-
-        <PrimaryActionCard item={item} />
-
-        {topAlertItems.length > 0 && (
-          <div className="mt-5 rounded-xl border border-slate-100 bg-slate-50/50 p-4">
-            <ConsolidatedEvidenceAlert key={item.request_id} items={topAlertItems} />
-          </div>
-        )}
-
-        {whatWouldHappen && (
-          <div className="mt-5">
-            <button
-              type="button"
-              onClick={handleToggleConsequences}
-              className="flex items-center gap-2 text-sm font-medium text-brand-blue focus:outline-none focus:ring-2 focus:ring-brand-blue/20 rounded-lg px-2 py-1 -ml-2"
-              aria-expanded={showConsequences}
-            >
-              <HiMiniInformationCircle className="h-4 w-4" aria-hidden="true" />
-              What would happen without Guard?
-              {showConsequences ? (
-                <HiMiniChevronUp className="h-3 w-3" aria-hidden="true" />
-              ) : (
-                <HiMiniChevronDown className="h-3 w-3" aria-hidden="true" />
-              )}
-            </button>
-            {showConsequences && (
-              <div className="mt-3 rounded-xl border border-slate-200/70 bg-slate-50 p-4">
-                <p className="text-sm text-brand-dark">{whatWouldHappen}</p>
-              </div>
-            )}
-          </div>
-        )}
-
-        <div className="mt-6 space-y-2">
-          <SectionLabel>Approval scope</SectionLabel>
-          {!hasAllowScope && (
-            <p className="text-sm text-brand-attention" role="status">
-              This action cannot be approved under its current Guard policy.
-            </p>
-          )}
-          <div className="grid grid-cols-1 gap-2 md:grid-cols-2" role="radiogroup" aria-label="Allow scope selection">
-            {commonScopeOptions.map((choice) => (
-              <ScopeChoiceButton
-                key={choice.value}
-                choice={choice}
-                checked={allowScope === choice.value}
-                onScopeChange={setAllowScope}
-              />
-            ))}
-          </div>
-          {broaderScopeOptions.length > 0 && (
-            <details className="rounded-xl border border-brand-blue/15 bg-brand-blue/[0.03] p-3">
-              <summary className="cursor-pointer select-none text-xs font-semibold uppercase tracking-[0.16em] text-brand-blue">
-                Save for project or app
-              </summary>
-              <p className="mt-2 text-xs text-brand-dark/70">
-                These options save a decision that skips review for matching actions going forward. Choose the narrowest scope that fits what you meant to allow.
-              </p>
-              <div className="mt-3 grid grid-cols-1 gap-2 md:grid-cols-2">
-                {broaderScopeOptions.map((choice) => (
-                  <ScopeChoiceButton
-                    key={choice.value}
-                    choice={choice}
-                    checked={allowScope === choice.value}
-                    onScopeChange={setAllowScope}
-                  />
-                ))}
-              </div>
-            </details>
-          )}
-          {advancedScopeOptions.length > 0 && (
-            <details className="rounded-xl border border-brand-attention/20 bg-brand-attention/[0.04] p-3">
-              <summary className="cursor-pointer select-none text-xs font-semibold uppercase tracking-[0.16em] text-brand-attention">
-                Advanced: save everywhere on this machine
-              </summary>
-              <p className="mt-2 text-xs text-brand-dark/70">
-                This saves a decision that applies across all your projects on this machine. Matching actions skip review permanently. Only use this if you fully trust this action everywhere.
-              </p>
-              <div className="mt-3 grid grid-cols-1 gap-2">
-                {advancedScopeOptions.map((choice) => (
-                  <ScopeChoiceButton
-                    key={choice.value}
-                    choice={choice}
-                    checked={allowScope === choice.value}
-                    onScopeChange={setAllowScope}
-                  />
-                ))}
-              </div>
-            </details>
-          )}
-          {taskCapabilityCopy !== null && (
-            <div className="flex items-start gap-2 pt-1 text-xs text-brand-dark/70">
-              <HiMiniKey className="mt-0.5 h-4 w-4 shrink-0 text-brand-blue" aria-hidden="true" />
-              <p>{taskCapabilityCopy}</p>
-            </div>
-          )}
-          {blockScopeOptions.length > 0 && (
-            <details className="rounded-xl border border-slate-200/70 bg-slate-50/60 p-3">
-              <summary className="cursor-pointer select-none text-xs font-semibold uppercase tracking-[0.16em] text-brand-dark">
-                Block matching actions
-              </summary>
-              <p className="mt-2 text-xs text-brand-dark/70">
-                Blocking can cover a wider trusted selector without granting permission to run anything.
-              </p>
-              <div className="mt-3 grid grid-cols-1 gap-2 md:grid-cols-2" role="radiogroup" aria-label="Block scope selection">
-                {blockScopeOptions.map((choice) => (
-                  <ScopeChoiceButton
-                    key={choice.value}
-                    choice={choice}
-                    checked={blockScope === choice.value}
-                    onScopeChange={setBlockScope}
-                  />
-                ))}
-              </div>
-            </details>
-          )}
-        </div>
-
-        {errorMessage && (
-          <div className="guard-fade-in mt-4 rounded-xl border border-brand-purple/25 bg-brand-purple/[0.05] p-4">
-            <div className="flex items-start gap-3">
-              <HiMiniExclamationTriangle className="mt-0.5 h-4 w-4 shrink-0 text-brand-purple" aria-hidden="true" />
-              <div className="flex-1">
-                <p className="text-sm text-brand-purple">{errorMessage}</p>
-                <button
-                  type="button"
-                  onClick={handleRetryLastAction}
-                  className="mt-2 inline-flex min-h-9 items-center rounded-lg border border-slate-200 bg-white px-3 text-sm font-medium text-brand-dark transition-colors hover:bg-slate-50"
-                >
-                  Retry
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
-
-        <div className="mt-6 grid grid-cols-1 gap-3 sm:grid-cols-2">
-          <ActionButton
-            ref={allowButtonRef}
-            variant="success"
-            onClick={handleAllow}
-            disabled={!hasAllowScope || submitting !== null || pendingAction !== null}
-          >
-            {submitting === "allow" ? (
-              <span className="flex items-center gap-2">
-                <HiMiniArrowPath className="h-4 w-4 animate-spin" aria-hidden="true" />
-                Approving...
-              </span>
-            ) : (
-              <span className="flex items-center gap-2">
-                <HiMiniCheckCircle className="h-4 w-4" aria-hidden="true" />
-                {allowButtonLabel(allowScope)}
-              </span>
-            )}
-          </ActionButton>
-          <ActionButton
-            variant="outline"
-            onClick={handleBlock}
-            disabled={blockScopeOptions.length === 0 || submitting !== null || pendingAction !== null}
-          >
-            {submitting === "block" ? (
-              <span className="flex items-center gap-2">
-                <HiMiniArrowPath className="h-4 w-4 animate-spin" aria-hidden="true" />
-                Blocking...
-              </span>
-            ) : (
-              <span className="flex items-center gap-2">
-                <HiMiniNoSymbol className="h-4 w-4" aria-hidden="true" />
-                {blockButtonLabel(blockScope)}
-              </span>
-            )}
-          </ActionButton>
-        </div>
-
-      </div>
-
-      {evidenceItems.length > 0 && (
-        <div className="rounded-xl border border-slate-100 p-4 sm:p-5">
-          <button
-            type="button"
-            onClick={handleToggleEvidence}
-            className="flex w-full items-center justify-between text-left focus:outline-none focus:ring-2 focus:ring-brand-blue/20 rounded-lg px-2 py-1 -ml-2"
-            aria-expanded={showEvidence}
-          >
-            <SectionLabel>Review details</SectionLabel>
-            {showEvidence ? (
-              <HiMiniChevronUp className="h-4 w-4 text-slate-400" aria-hidden="true" />
-            ) : (
-              <HiMiniChevronDown className="h-4 w-4 text-slate-400" aria-hidden="true" />
-            )}
-          </button>
-          {showEvidence && (
-            <div className="mt-4">
-              <ConsolidatedEvidenceAlert key={item.request_id} items={evidenceItems} />
-            </div>
-          )}
-        </div>
-      )}
-
-      {detail.receipt && (
-        <div className="rounded-xl border border-slate-100 p-4 sm:p-5">
-          <SectionLabel>Last time</SectionLabel>
-          <p className="mt-2 text-sm text-muted-foreground">
-            You previously {pastDecisionVerb(detail.receipt.policy_decision)} a similar action{" "}
-            {formatRelativeTime(detail.receipt.timestamp)}.
-          </p>
-          {detail.diff && detail.diff.changed_fields.length > 0 && (
-            <div className="mt-3 rounded-xl border border-slate-200/70 bg-slate-50 p-4">
-              <p className="text-sm font-medium text-brand-dark">What changed since then:</p>
-              <ul className="mt-2 space-y-1">
-                {detail.diff.changed_fields.map((field) => (
-                  <li key={field} className="flex items-center gap-2 text-sm text-brand-dark">
-                    <HiMiniCheckCircle className="h-3.5 w-3.5 shrink-0 text-brand-blue" aria-hidden="true" />
-                    {field}
-                  </li>
-                ))}
-              </ul>
-            </div>
-          )}
-        </div>
-      )}
-
-      {pendingAction !== null && props.approvalGate !== null && (
-        <ApprovalPasswordModal
-          gate={props.approvalGate}
-          approvalPassword={approvalPassword}
-          approvalTotpCode={approvalTotpCode}
-          useCooldown={useCooldown}
-          onApprovalPasswordChange={handleApprovalPasswordChange}
-          onApprovalTotpCodeChange={handleApprovalTotpCodeChange}
-          onUseCooldownChange={handleUseCooldownChange}
-          onSubmit={handleModalSubmit}
-          onCancel={handleModalCancel}
-          submitLabel={pendingAction === "allow" ? allowButtonLabel(allowScope) : blockButtonLabel(blockScope)}
-        />
-      )}
-    </div>
-  );
-}
-
-function ScopeChoiceButton(props: {
-  choice: ApprovalScopeChoice;
-  checked: boolean;
-  onScopeChange: (scope: DecisionScope) => void;
-}) {
-  const handleClick = useCallback(() => {
-    props.onScopeChange(props.choice.value);
-  }, [props.onScopeChange, props.choice.value]);
-
-  return (
-    <button
-      type="button"
-      onClick={handleClick}
-      role="radio"
-      aria-checked={props.checked}
-      className={`rounded-xl border px-4 py-3 text-left transition-all focus:outline-none focus:ring-2 focus:ring-brand-blue/20 ${
-        props.checked ? "border-brand-blue bg-brand-blue/[0.06]" : "border-slate-200/70 bg-white hover:bg-slate-50"
-      }`}
-    >
-      <p className="text-sm font-medium text-brand-dark">{props.choice.label}</p>
-      <p className="mt-0.5 text-xs text-muted-foreground">{props.choice.description}</p>
-    </button>
-  );
-}
-
-function allowButtonLabel(scope: DecisionScope): string {
-  if (scope === "artifact") {
-    return "Approve once";
-  }
-  if (scope === "workspace") {
-    return "Remember for project";
-  }
-  return "Approve and remember";
-}
-
-function blockButtonLabel(scope: DecisionScope): string {
-  if (scope === "artifact") {
-    return "Keep blocked";
-  }
-  if (scope === "workspace") {
-    return "Block in project";
-  }
-  return "Block matching actions";
-}
-
-type ReviewCodexResumePanelProps = {
-  resume: GuardCodexResumeResult;
-  onRetry?: () => void;
-};
-
-function ReviewCodexResumePanel({ resume, onRetry }: ReviewCodexResumePanelProps) {
-  const ux = buildCodexResumeUx(resume);
-  const isPending = resume.status === "pending" || resume.status === "in_progress";
-  const isSuccess = resume.status === "sent" || resume.status === "already_sent";
-  const isFailed = resume.status === "failed";
-
-  const borderClass = isFailed
-    ? "border-brand-purple/25 bg-brand-purple/[0.05]"
-    : isSuccess
-    ? "border-brand-green/25 bg-brand-green-bg/30"
-    : isPending
-    ? "border-brand-blue/25 bg-brand-blue/[0.04]"
-    : "border-slate-200/60 bg-slate-50/40";
-
-  const iconClass = isFailed
-    ? "text-brand-purple"
-    : isSuccess
-    ? "text-brand-green"
-    : "text-brand-blue";
-
-  return (
-    <div className={`flex items-start gap-3 rounded-2xl border px-4 py-3 ${borderClass}`}>
-      {isPending && (
-        <HiMiniArrowPath className={`mt-0.5 h-4 w-4 shrink-0 animate-spin ${iconClass}`} aria-hidden="true" />
-      )}
-      {isSuccess && (
-        <HiMiniCheckCircle className={`mt-0.5 h-4 w-4 shrink-0 ${iconClass}`} aria-hidden="true" />
-      )}
-      {isFailed && (
-        <HiMiniExclamationTriangle className={`mt-0.5 h-4 w-4 shrink-0 ${iconClass}`} aria-hidden="true" />
-      )}
-      {!isPending && !isSuccess && !isFailed && (
-        <HiMiniInformationCircle className="mt-0.5 h-4 w-4 shrink-0 text-slate-500" aria-hidden="true" />
-      )}
-      <div className="flex-1 space-y-1">
-        <p className="text-sm font-medium text-brand-dark">{ux.headline}</p>
-        {ux.body !== null && (
-          <p className="text-xs text-muted-foreground">{ux.body}</p>
-        )}
-        {isFailed && onRetry !== undefined && (
-          <div className="mt-2">
-            <ActionButton variant="outline" onClick={onRetry}>
-              Retry resume
-            </ActionButton>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function ReviewEmptyState({ runtime, resolutionMessage, codexResume, onRetryResume }: { runtime: GuardRuntimeSnapshot | null; resolutionMessage: string | null; codexResume: GuardCodexResumeResult | null; onRetryResume?: () => void }) {
-  const protectionHealth = runtime ? protectionHealthFor(runtime) : unavailableProtectionHealth();
-  const protectedAppsCount = protectionHealth.apps.filter((app) => app.state === "protected").length;
-  const heroStatus = protectionHealth.state === "protected" ? "clear" : protectionHealth.state;
-  const {
-    Icon: ProtectionIcon,
-    cardClass: healthCardClass,
-    iconClass: healthIconClass,
-  } = PROTECTION_APPEARANCE[protectionHealth.state];
-
-  return (
-    <div className="space-y-6">
-      <GuardHero
-        status={heroStatus}
-        headline="Nothing to review"
-        subheadline={`No actions need your decision right now. ${protectionHealth.detail}`}
-      />
-
-      <ProofStrip
-        items={[
-          { label: "Queue", value: "All clear", tone: "green" },
-          { label: "Protection", value: protectionHealth.label, tone: protectionHealth.state === "protected" ? "green" : "slate" },
-          { label: "Apps protected", value: protectedAppsCount, tone: protectedAppsCount > 0 ? "green" : "slate" },
-        ]}
-      />
-
-      {codexResume !== null && (
-        <ReviewCodexResumePanel resume={codexResume} onRetry={onRetryResume} />
-      )}
-
-      {codexResume === null && resolutionMessage && (
-        <div className="flex items-start gap-3 rounded-2xl border border-brand-green/25 bg-brand-green-bg/30 px-4 py-3">
-          <HiMiniCheckCircle className="mt-0.5 h-4 w-4 shrink-0 text-brand-green" aria-hidden="true" />
-          <p className="text-sm font-medium text-brand-green-text">{resolutionMessage}</p>
-        </div>
-      )}
-
-      <div className="grid gap-6 lg:grid-cols-2">
-        <div className={`rounded-xl border p-4 sm:p-5 ${healthCardClass}`}>
-          <div className="flex items-start gap-3">
-            <span className={`inline-flex h-10 w-10 shrink-0 items-center justify-center rounded-full ${healthIconClass}`}>
-              <ProtectionIcon className="h-5 w-5" aria-hidden="true" />
-            </span>
-            <div>
-              <SectionLabel>{protectionHealth.label}</SectionLabel>
-              <p className="mt-2 text-sm text-muted-foreground">
-                {protectionHealth.detail} When something needs review, it will appear here.
-              </p>
-            </div>
-          </div>
-        </div>
-
-        <div className="rounded-xl border border-slate-100 p-4 sm:p-5">
-          <SectionLabel>What Guard does</SectionLabel>
-          <ul className="mt-3 space-y-2">
-            {[
-              "Pauses risky file reads and writes",
-              "Blocks commands that could delete data",
-              "Warns about new network connections",
-              "Stops credential sharing",
-            ].map((item) => (
-              <li key={item} className="flex items-start gap-2 text-sm text-brand-dark">
-                <HiMiniCheckCircle className="mt-0.5 h-3.5 w-3.5 shrink-0 text-brand-green" aria-hidden="true" />
-                {item}
-              </li>
-            ))}
-          </ul>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function PrimaryActionCard({ item }: { item: GuardApprovalRequest }) {
-  const action = buildPrimaryReviewAction(item);
-
-  return (
-    <div className="mt-5 rounded-2xl border border-slate-200 bg-white p-4 shadow-sm">
-      <div className="flex flex-wrap items-center justify-between gap-3">
-        <div>
-          <SectionLabel>What was stopped</SectionLabel>
-          {action.detail !== null && (
-            <p className="mt-1 text-sm text-brand-dark/70">
-              {action.detail}
-            </p>
-          )}
-        </div>
-        <span className="rounded-full border border-brand-blue/15 bg-brand-blue/[0.04] px-3 py-1 font-mono text-[11px] font-semibold uppercase tracking-[0.16em] text-brand-blue">
-          {action.label}
-        </span>
-      </div>
-      <div className="mt-3">
-        <LoggedActionPanel
-          key={item.request_id}
-          label={action.label}
-          text={action.text}
-          copyAriaLabel="Copy full stopped action to clipboard"
-          expandAriaLabel="Expand full stopped action"
-          collapseAriaLabel="Collapse full stopped action"
-        />
-      </div>
-    </div>
-  );
-}
-
-function buildWhatWouldHappen(item: GuardApprovalRequest): string | null {
-  const type = item.artifact_type;
-  if (type?.includes("file_write") || type?.includes("file_read")) {
-    return `Without Guard, ${harnessDisplayName(item.harness)} would access "${item.artifact_name ?? item.artifact_id}" immediately. Guard paused it so you can review first.`;
-  }
-  if (type?.includes("shell") || type?.includes("command")) {
-    return `Without Guard, this shell command would run immediately. Guard paused it so you can review what it does first.`;
-  }
-  if (type?.includes("network") || type?.includes("request")) {
-    return `Without Guard, this request would go to the network immediately. Guard paused it so you can review the destination first.`;
-  }
-  if (type?.includes("mcp") || type?.includes("tool")) {
-    return `Without Guard, this tool would execute immediately. Guard paused it so you can review what data it accesses.`;
-  }
-  return `Without Guard, this action would run immediately. Guard paused it so you can review and decide.`;
-}
-
-function pastDecisionVerb(decision: string): string {
-  if (decision === "allow") {
-    return "allowed";
-  }
-  if (decision === "block") {
-    return "blocked";
-  }
-  return "reviewed";
 }
